添加区域id在院校机构上有显示，就不能删除

+++ src/main/java/com/bc/pmpheep/back/service/AreaServiceImpl.java
@@ -80,6 +80,10 @@ public class AreaServiceImpl extends BaseService implements AreaService {
 		if (null == id) {
 			throw new CheckedServiceException(CheckedExceptionBusiness.AREA, CheckedExceptionResult.NULL_PARAM, "主键为空");
 		}
+		if (1 < areaDao.getOrgId(id)) {
+			throw new CheckedServiceException(CheckedExceptionBusiness.AREA, CheckedExceptionResult.NULL_PARAM, 
+					"区域在机构区域已使用");
+		}
 		return areaDao.deleteAreaById(id);
 	}
